main関数の分離と、sortのヘッダーファイル作成
* ソート異常が存在する。
